repro: make the release gates clean-checkout reproducible + fusd-math strict clippy
External audit (2026-07-11) remediation item 6. A fresh clone can now run the
full gate suite unaided:

- check-no-dev-oracle.sh asserts the built IDL address == the declared mainnet
  program id, so the scanned bytecode is provably built from canonical source
  (and an accidental 'anchor keys sync' — which rewrites declare_id! to the
  random throwaway keypair anchor generates on a clean clone — is caught
  instead of silently invalidating the scan). The gates never read the deploy
  keypair; scripts/README.md's misleading claim that target/deploy must match
  declare_id! is rewritten to say exactly that.
- The litesvm harness .so load now fails with an actionable build instruction
  on a clean checkout, and rejects the PRODUCTION .so that ci-checks.sh step 7
  deliberately leaves behind (detected via the 'Instruction: DevSetPrice'
  codegen marker — the same string check-no-dev-oracle.sh asserts is absent).
- README gains the clean-checkout note (no keys needed; bare 'cargo test
  --workspace' needs the dev-oracle build first).
- Strict clippy extends to fusd-math (ci-checks step 2): fixes the 2 lib lints
  the audit found (needless_range_loop in rate_bucket::first_set_from — now the
  index-identical iterator form; doc_lazy_continuation in recovery.rs) plus the
  5 test-target lints cargo's fail-fast masked (unnecessary_cast,
  3x manual_range_contains, manual_is_multiple_of; needs host Rust >= 1.87).
  124 fusd-math tests green; clippy -D warnings exits 0.

+++ crates/fusd-math/src/interest.rs
@@ -192,7 +192,7 @@ mod tests {
         };
         for _ in 0..20_000 {
             // debt up to ~1e15 native ($1B), rate up to 2550 bps, period up to ~20 years.
-            let debt = (next() % 1_000_000_000_000_000) as u128;
+            let debt = next() % 1_000_000_000_000_000;
             let rate = (next() % 2_551) as u16;
             let period = (next() % 631_152_000) as u64; // 0..~20y
